+++ src/main/java/com/mobble/mobbleserver/refactor/meeting/service/MeetingService.java
@@ -3,9 +3,9 @@ package com.mobble.mobbleserver.refactor.meeting.service;
 import com.mobble.mobbleserver.refactor.club.policy.ClubPermissionPolicy;
 import com.mobble.mobbleserver.refactor.clubMember.entity.ClubMember;
 import com.mobble.mobbleserver.refactor.clubMember.validator.ClubMemberValidator;
-import com.mobble.mobbleserver.refactor.meeting.dto.request.MeetingRequestDto;
-import com.mobble.mobbleserver.refactor.meeting.dto.request.MeetingUpdateRequestDto;
-import com.mobble.mobbleserver.refactor.meeting.dto.response.MeetingResponseDto;
+import com.mobble.mobbleserver.infrastructure.web.meeting.dto.request.MeetingRequestDto;
+import com.mobble.mobbleserver.infrastructure.web.meeting.dto.request.MeetingUpdateRequestDto;
+import com.mobble.mobbleserver.infrastructure.web.meeting.dto.response.MeetingResponseDto;
 import com.mobble.mobbleserver.infrastructure.web.meeting.Meeting;
 import com.mobble.mobbleserver.refactor.meeting.repository.MeetingRepository;
 import com.mobble.mobbleserver.refactor.meeting.validator.MeetingValidator;
